Removed result count from header
Adding the result count to the header without handling the model being null causes the page to error out when the app is first started. Removing the result count from the header (and leaving it only alongside the results within the Model != null block as outlined in the demo docs) resolves this issue.